+++ generic_u1/io_u1lat.c
@@ -1512,7 +1512,7 @@ int read_u1_gauge_hdr(gauge_file *gf, int parallel)
 	  ny = gh->dims[1];
 	  nz = gh->dims[2];
 	  nt = gh->dims[3];
-	  volume = nx*ny*nz*nt;
+	  volume = (size_t)nx*ny*nz*nt;
 	}
     }
